Major update to the URL service code.
The url service code (ServiceCode.pys) has been updated so that it now plays back HLS streams instead of flash-based webkit streams. HLS is much more stable and reliable. It is also much better quality. This is possible because i figured out how to pass the session_id created by the main plugin code to the service code. I am doing this by adding it onto the end of the URL being sent to the service code. Once the URL is in the service code, the session_id is stripped of the URL. This session_id allows me to make API calls to CR to get access to the HLS streams. This is a bit hacky but is the only way to do it due to Plex's limitations. Hopefully in the future Plex will allow the service code to access the main plugin's DICT so i can access the session_id from the main plugin code properly. (I've heard rumors that a future PMS update will do this)

The only update to the main plugin code (__init__.py) was to add the session_id on to the end of the URL that is being sent to the service code.

Lastly, this update is the first of several more. I will be updating the main plugin code to only allow premium users access. This is a Crunchyroll limitation and i don't see any way around it. If you do not have a premium account, the HLS streams don't work correctly. Therefor, free accounts won't work with this plugin. This also follows Crunchyroll's model of only allowing premium members access to their apps for the various platforms they support. (PS3, XBox, Roku, etc)